Refactor: move db-utils to `util/db.ts`
