+++ src/commands/stat.ts
@@ -1,6 +1,6 @@
 import { Command } from '../types/command.js';
 import { EmbedBuilder } from '@discordjs/builders';
-import { upsertUser } from '../util/upsertUser.js';
+import { upsertUser } from '../util/db.js';
 import { PrismaClient } from '@prisma/client';
 import { getScore } from '../util/score.js';
